+++ asyncband/src/mpsc/bounded/buffer.rs
@@ -393,10 +393,10 @@ mod tests {
         value: T,
     ) -> Result<(), T> {
         // SAFETY: The test claimed this position while holding the same capacity permit.
-        unsafe { tx.shared().buffer.slots().publish(position, value) }?;
+        unsafe { tx.shared.buffer.slots().publish(position, value) }?;
         // Publication owns the capacity now; forgetting skips the permit's release on drop.
         std::mem::forget(permit);
-        tx.shared().rx_waker.wake();
+        tx.shared.rx_waker.wake();
         Ok(())
     }
 
@@ -406,7 +406,7 @@ mod tests {
             for initial in [0, usize::MAX - 1] {
                 let (tx, mut rx) = bounded(capacity);
                 // Start an empty ring near ticket overflow instead of running usize::MAX sends.
-                tx.shared()
+                tx.shared
                     .buffer
                     .slots()
                     .tail
@@ -415,7 +415,7 @@ mod tests {
                 let mut cx = Context::from_waker(Waker::noop());
                 for lap in 0..8 {
                     let permit = tx.try_reserve().unwrap();
-                    let position = tx.shared().buffer.slots().claim().unwrap();
+                    let position = tx.shared.buffer.slots().claim().unwrap();
                     for offset in 1..capacity {
                         tx.try_send(lap * capacity + offset).unwrap();
                     }
@@ -441,16 +441,11 @@ mod tests {
             drops: drops.clone(),
             _sender: tx.clone(),
         };
-        let allocation = Arc::downgrade(tx.shared());
+        let allocation = Arc::downgrade(&tx.shared);
         // Pause after claim's open check, then resume its atomic ticket allocation after close.
-        assert!(!tx.shared().buffer.slots().closed.load(Ordering::Acquire));
+        assert!(!&tx.shared.buffer.slots().closed.load(Ordering::Acquire));
         drop(rx);
-        let position = tx
-            .shared()
-            .buffer
-            .slots()
-            .tail
-            .fetch_add(1, Ordering::AcqRel);
+        let position = tx.shared.buffer.slots().tail.fetch_add(1, Ordering::AcqRel);
         let unsent = publish_claimed(&tx, permit, position, value).unwrap_err();
         assert_eq!(unsent.bytes, [7; 1024]);
         drop(unsent);
@@ -477,7 +472,7 @@ mod tests {
     #[test]
     fn closing_reclaims_ready_values_without_waiting_for_a_paused_publisher() {
         let (tx, rx) = bounded(2);
-        let allocation = Arc::downgrade(tx.shared());
+        let allocation = Arc::downgrade(&tx.shared);
         let drops = Arc::new(AtomicUsize::new(0));
         let paused = Barrier::new(2);
         let (resume_tx, resume_rx) = std::sync::mpsc::channel();
@@ -489,7 +484,7 @@ mod tests {
             let paused = &paused;
             let publisher = scope.spawn(move || {
                 let permit = sender.try_reserve().unwrap();
-                let position = sender.shared().buffer.slots().claim().unwrap();
+                let position = sender.shared.buffer.slots().claim().unwrap();
                 let value = Payload {
                     bytes: [1; 1024],
                     drops: drops.clone(),
@@ -534,7 +529,7 @@ mod tests {
     fn publication_racing_with_close_drops_every_payload_once() {
         for _ in 0..if cfg!(miri) { 8 } else { 128 } {
             let (tx, rx) = bounded(3);
-            let allocation = Arc::downgrade(tx.shared());
+            let allocation = Arc::downgrade(&tx.shared);
             let drops = Arc::new(AtomicUsize::new(0));
             let start = Barrier::new(4);
             thread::scope(|scope| {
