+++ asyncband/src/mpsc/bounded/mod.rs
@@ -35,10 +35,6 @@ mod receiver;
 mod semaphore;
 mod sender;
 
-pub use self::receiver::BoundedReceiver;
-pub use self::sender::BoundedSender;
-pub use self::sender::Permit;
-
 /// Creates a bounded mpsc channel with room for `buffer` queued messages.
 ///
 /// [`BoundedSender::send`] waits for capacity when the buffer is full. Receiving a message releases
@@ -53,32 +49,65 @@ pub use self::sender::Permit;
 /// Panics if `buffer` is zero or exceeds the maximum capacity of `usize::MAX >> 1`.
 #[track_caller]
 pub fn bounded<T>(buffer: usize) -> (BoundedSender<T>, BoundedReceiver<T>) {
-    assert!(buffer > 0, "mpsc bounded channel requires buffer > 0");
+    /// The largest capacity accepted by [`bounded`].
+    ///
+    /// The shared permit counter reserves two sentinel values above the usable range, and the
+    /// zero-sized queue length packs a closed flag into its top bit. This bound also keeps the
+    /// rounded-up slot storage from overflowing a power of two.
+    const MAX_CAPACITY: usize = usize::MAX >> 1;
+
+    assert!(
+        buffer > 0,
+        "mpsc bounded channel capacity {buffer} must be nonzero",
+    );
     assert!(
         buffer <= MAX_CAPACITY,
-        "mpsc bounded channel capacity {buffer} exceeds the maximum of {MAX_CAPACITY}"
+        "mpsc bounded channel capacity {buffer} exceeds the maximum of {MAX_CAPACITY}",
     );
+
     let shared = Arc::new(Shared {
         senders: AtomicUsize::new(1),
         tx_permits: CachePadded::new(Semaphore::new(buffer)),
         rx_waker: AtomicWaker::new(),
         buffer: Buffer::new(buffer),
     });
-    let sender = BoundedSender::new(shared.clone());
-    let receiver = BoundedReceiver::new(shared);
+    let sender = BoundedSender {
+        shared: shared.clone(),
+    };
+    let receiver = BoundedReceiver { shared, head: 0 };
     (sender, receiver)
 }
 
-pub struct Shared<T> {
+/// The sending endpoint of a bounded mpsc channel.
+///
+/// Instances are created by the [`bounded`] function.
+pub struct BoundedSender<T> {
+    shared: Arc<Shared<T>>,
+}
+
+/// The receiving endpoint of a bounded mpsc channel.
+///
+/// Instances are created by the [`bounded`] function. Dropping the receiver discards queued values.
+/// The backing allocation remains alive until all endpoints are dropped, so a concurrent sender
+/// can safely finish returning an unsent value.
+pub struct BoundedReceiver<T> {
+    shared: Arc<Shared<T>>,
+    head: usize,
+}
+
+/// Capacity reserved for one message on a bounded channel.
+///
+/// Created by [`BoundedSender::reserve`] or [`BoundedSender::try_reserve`]. Holding a permit
+/// reduces available capacity but does not prevent other messages from being received. Dropping
+/// it without sending releases capacity and notifies a waiting sender.
+#[must_use = "dropping the permit releases its reserved capacity"]
+pub struct Permit<'a, T> {
+    sender: Option<&'a BoundedSender<T>>,
+}
+
+struct Shared<T> {
     senders: AtomicUsize,
     tx_permits: CachePadded<Semaphore>,
     rx_waker: AtomicWaker,
     buffer: Buffer<T>,
 }
-
-/// The largest capacity accepted by [`bounded`].
-///
-/// The shared permit counter reserves two sentinel values above the usable range, and the
-/// zero-sized queue length packs a closed flag into its top bit. This bound also keeps the
-/// rounded-up slot storage from overflowing a power of two.
-const MAX_CAPACITY: usize = usize::MAX >> 1;
